refactor: merge authentication_context module into auth.py
Consolidate the ambient-authentication primitive into auth.py where it
lives naturally alongside Authentication, AuthHeader, NoAuth, and the
existing auth registry. Delete the standalone authentication_context.py
(was 70 lines) — there is no meaningful "one concern per module" split
between the authentication types and the ambient scope machinery that
references them.

Merge the corresponding test_authentication_context.py tests into
test_auth.py so test layout mirrors source layout. The existing
clean_auth_registry autouse fixture is harmless for the new tests.

celeste.__init__ and __all__ now import AuthenticationContext,
authentication_scope, and resolve_authentication directly from
celeste.auth. Public API surface is unchanged — top-level imports still
work via `from celeste import authentication_scope, AuthenticationContext`.
